refactor: address P2 review findings across backend and iOS
- Extract dual-auth logic (Bearer token / Clerk) into shared
  lib/resolve-user-id.ts utility
- Move auth check before body parsing in item/create to reject
  unauthenticated requests earlier
- Wrap user upsert + token upsert in prisma.$transaction for atomicity
- Add composite index on (userId, url, isDeleted) for duplicate checks
- Deduplicate iOS constants into shared Constants.swift (app group ID,
  default server URL, shared UserDefaults)
