+++ app/api/item/create/route.ts
@@ -1,30 +1,18 @@
 "use server";
 
-import { auth } from "@clerk/nextjs/server";
-import crypto from "crypto";
-import { headers } from "next/headers";
 import { NextResponse } from "next/server";
 
 import { addTwitterPostOrBookmark } from "@/lib/add-twitter-post-or-bookmark";
 import { addWebsite } from "@/lib/add-website";
 import { checkDuplicateItem } from "@/lib/check-duplicate-item";
-import prisma from "@/lib/prisma";
+import { resolveUserId } from "@/lib/resolve-user-id";
 import { isTwitterPostOrBookmarkUrl, normalizeLink } from "@/lib/url";
 
 export async function POST(req: Request) {
-  let userId: string | null = null;
-  const authorization = headers().get("authorization");
+  const userId = await resolveUserId();
 
-  if (authorization?.startsWith("Bearer coolection_")) {
-    const token = authorization.substring(7);
-    const tokenHash = crypto.createHash("sha256").update(token).digest("hex");
-    const apiToken = await prisma.apiToken.findUnique({
-      where: { tokenHash },
-      select: { userId: true },
-    });
-    userId = apiToken?.userId ?? null;
-  } else {
-    userId = auth().userId;
+  if (!userId) {
+    return new NextResponse("Unauthorized", { status: 401 });
   }
 
   const body = await req.json();
@@ -32,10 +20,6 @@ export async function POST(req: Request) {
 
   if (!url) {
     return NextResponse.json({ message: "URL is required" }, { status: 400 });
-  }
-
-  if (!userId) {
-    return new NextResponse("Unauthorized", { status: 401 });
   }
 
   const normalizedLink = normalizeLink(url);
