+++ ios/CoolectionSafari/Shared/Constants.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+enum AppConstants {
+    static let appGroup = "group.coolection.app"
+    static let defaultServer = "https://coolection.co"
+    static let defaults = UserDefaults(suiteName: appGroup)
+}
